ex3.js

+++ ex3.js
@@ -1,7 +1,7 @@
 //Exercício 3: Ímpar ou par
 //Faça um programa que receba um número do usuário e informe se este número é par ou ímpar.
 
-let numero = parseFloat(prompt("Escreva qualquer numero:"))
+let numero = parseInt(prompt("Escreva qualquer numero:"))
 
 if (numero % 2 == 0) {
     console.log ("O número Par")
